fix(task3): Correct missing item in the database

+++ UniversityAPI/UniversityAPI/Services/DatabaseService.cs
@@ -8,12 +8,13 @@ namespace UniversityAPI.Services;
 public class DatabaseService
 {
     private readonly string dbPath = @"Data/students.csv";
+
     // INFO: This is a design choice of when the application startup I read the file from disk and keep it in memory.
     // This is due to simplicity and I don't expect file being updated externally.
     // In a closer to real life problem I would for sure cache the information to avoid I/O in every call but would keep
     // reading from disk from time to time or simply read it all the time using ASYNC methods.
     private List<Student> db;
-    
+
     public DatabaseService()
     {
         db = CSVParser.Parse(dbPath);
@@ -29,14 +30,13 @@ public class DatabaseService
         var index = db.FindIndex(student => student.IndexNumber == indexNumber);
 
         return index;
-
     }
-    
+
     public Student? Get(string indexNumber)
     {
         var index = GetIndex(indexNumber);
 
-        return db[index];
+        return index > -1 ? db[index] : null;
     }
 
     public void Create(Student newRecord)
@@ -49,8 +49,10 @@ public class DatabaseService
     {
         var index = GetIndex(record.IndexNumber);
 
-        db[index] = record;
+        if (index == -1) return;
         
+        db[index] = record;
+
         Flush();
     }
 
@@ -58,8 +60,10 @@ public class DatabaseService
     {
         var index = GetIndex(indexNumber);
         
+        if (index == -1) return;
+
         db.RemoveAt(index);
-        
+
         Flush();
     }
 
@@ -72,7 +76,7 @@ public class DatabaseService
             // just for simplicity
             sb.AppendLine(record.ToString());
         }
-        
+
         File.WriteAllText(dbPath, sb.ToString());
     }
 }
